Split the huge mp file into multiple smaller files.
